+++ mzLib/Test/Omics/Occupancy/ModificationOccupancyCalculatorTests.cs
@@ -7,7 +7,7 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 
-namespace Test.Omics;
+namespace Test.Omics.Occupancy;
 
 [TestFixture]
 [ExcludeFromCodeCoverage]
